UPDATE ITENS A SEREM SUBSTITUIDOS

+++ Service/Dashboards/Produtividades.py
@@ -47,7 +47,7 @@ def ProdutividadeGarantiaEquipe(dataInico, dataFim , horaInicio, horaFim):
 
     conn.close()
     consulta['qtd'].fillna(0,inplace=True)
-    consulta['qtd'] = consulta['qtd'].astype(int)
+    consulta['qtd'] = consulta['qtd'].astype(float)
     consulta  = consulta.groupby(['operador1','operador2','operador3'])['qtd'].sum().reset_index()
